fix(omnivoice): silence -Werror=double-promotion in voice classifiers (#26)
The ubuntu-24-llguidance CI job builds the fused omnivoice tool with
-DLLAMA_FATAL_WARNINGS=ON, which turns -Wdouble-promotion into an error. Five
intentional float→double accumulations in the scalar voice-classifier forwards
(voice_speaker.c DC-removal, mel mean-centering, TDNN stats-pool mean;
voice_diarizer.c instance-norm mean/var) promoted a float operand implicitly.
Make each promotion explicit with a (double) cast — identical numerics, no more
warning. The sibling conv/linear accumulators are float and correctly untouched.

+++ tools/omnivoice/src/voice-classifiers/voice_classifier/voice_diarizer.c
@@ -114,11 +114,11 @@ static void inst_norm_1d_inplace(float *buf, int C, int L,
     for (int c = 0; c < C; ++c) {
         float *row = buf + (size_t)c * L;
         double sum = 0.0, sq = 0.0;
-        for (int i = 0; i < L; ++i) sum += row[i];
+        for (int i = 0; i < L; ++i) sum += (double)row[i];
         const float mean = (float)(sum / L);
         for (int i = 0; i < L; ++i) {
             const float d = row[i] - mean;
-            sq += (double)d * d;
+            sq += (double)d * (double)d;
         }
         const float var = (float)(sq / L);
         const float invstd = 1.0f / sqrtf(var + eps);
